Fix 3 things from previous iterations
1/ When users input the second number right after
choosing the operator, the number will appear on
the display. Before, the display won't show anything.

This is solved by adding an event listener to just the
first digit of the second number. Once users start
inputting the subsequent digits, the program will
remove that event listener and create a new one for
those.

2/ Before, users have to click "=" to get the final
result of an operation. Now, they can click any of the
4 operators and the program will show the result. It will
also remember that result for the next operation.

This is solved by adding a separate event listener for the
operators after users click the first time, namely '2nd operator'.
Of course, the event listener for the first time will be removed.

3/ When users click the "=" button, the program will remove
the event listener for the '2nd operator' and add a new one to
the operators as if it is the first time users click on them.
This is because once users click the "=" button, that result
becomes the first input number. There is no second input number
YET so clicking any operators won't do anything. The program
needs users to input the second number before proceeding.