Add function to the "." button
This allows users to input a decimal point if
they wish to do math with floating point numbers.
Users can only input "." once. The button will
be disabled if there's already a "." on the
screen.

There is also a special case where users don't input
the integer portion, which is assumed to be 0 at that
point. The program will take care of that as well.