Modify function for the "=" or solution button
Before, users can't start a brand new operation after
clicking the "=" button. If they input any digit, it
will be added to the existing number on the screen.
And that number will be used as the first number for
the next operation.

Now, if users don't want to reuse the result from
the previous operation, they can input any digit
and that new number will be on the screen. The program
will automatically erase the previous result and start
a new operation.